💥 Unifi deploy and serve behaviour
Migration Guide:

When running `deploy` the JS file will now be the same as with `serve`, `dist/main.js` instead of `dist/main.min.js`. Configure your build system to output to the same file for both development and production builds.

When running `serve` the HTML file will be minimised, this is to match the `deploy` behaviour.